+++ test/src/_paths.dart
@@ -2,6 +2,7 @@ const //
     allSourcesDirsPath = 'test/resources/samples/all_sources_dirs',
     allSourcesDirsMultiPath = 'test/resources/samples/all_sources_dirs_multi',
     emptyYamlPath = 'test/resources/samples/empty_yaml',
+    meantForAddingPath = 'test/resources/samples/meant_for_adding',
     meantForFixingPath = 'test/resources/samples/meant_for_fixing',
     meantForFixingEmptyPath = 'test/resources/samples/meant_for_fixing_empty',
     noDependenciesPath = 'test/resources/samples/no_dependencies',